+++ packages/pros-sys/src/lib.rs
@@ -14,12 +14,12 @@ pub mod ext_adi;
 pub mod gps;
 pub mod imu;
 pub mod link;
-pub mod llemu;
 pub mod misc;
 pub mod motor;
 pub mod optical;
 pub mod rotation;
 pub mod rtos;
+pub mod screen;
 pub mod vision;
 
 use core::ffi::{c_char, c_int, c_void};
@@ -32,12 +32,12 @@ pub use ext_adi::*;
 pub use gps::*;
 pub use imu::*;
 pub use link::*;
-pub use llemu::*;
 pub use misc::*;
 pub use motor::*;
 pub use optical::*;
 pub use rotation::*;
 pub use rtos::*;
+pub use screen::*;
 #[cfg(feaute = "apix")]
 pub use serial::*;
 pub use vision::*;
@@ -56,4 +56,35 @@ extern "C" {
     pub fn puts(s: *const c_char) -> i32;
     pub fn exit(code: i32) -> !;
     pub fn write(fd: c_int, buf: *const c_void, count: usize) -> isize;
+
+    fn initialize();
+    fn opcontrol();
+    fn autonomous();
+    fn disabled();
+    fn competition_initialize();
+}
+
+#[no_mangle]
+unsafe extern "C" fn cpp_opcontrol() {
+    opcontrol();
+}
+#[no_mangle]
+unsafe extern "C" fn cpp_autonomous() {
+    autonomous();
+}
+#[no_mangle]
+unsafe extern "C" fn cpp_disabled() {
+    disabled();
+}
+#[no_mangle]
+unsafe extern "C" fn cpp_competition_initialize() {
+    competition_initialize();
+}
+#[no_mangle]
+unsafe extern "C" fn cpp_initialize() {
+    initialize();
+}
+#[no_mangle]
+unsafe extern "C" fn task_fn_wrapper(function: task_fn_t, args: *mut c_void) {
+    function.unwrap()(args);
 }