+++ CHANGELOG.md
@@ -26,15 +26,27 @@ Before releasing:
 - Lightly document all APIs with missing documentation. (#70)
 - Added `Debug`, `Copy`, and `Clone` derives for common structs (#70)
 
+- Screen drawing API. (#81)
+- Added screen field to `Peripherals` and `DynamicPeripherals::take_screen` method. (#81)
+
 ### Fixed
 
 ### Changed
 
 - Re-exported printing macros from `pros::io`. (#82)
 - Applied several lints to improve code quality. (#70)
+- Updated to PROS version 4. (**Breaking Change**) (#81)
+- Moved `vision::Rgb` into its own `color.rs` file. (**Breaking Change**) (#81)
+- The VEXOS target has been updated to improve file size and floating point operation speed. (#81)
+- `Peripherals::new()` is no longer const (**Breaking Change) (#81)
+- Updated panic handler to print to the brain display as well as over serial (#81)
 
 ### Removed
 
+- LVGL bindings (pros-sys) and colors (pros). (**Breaking Change**) (#81)
+- LLEMU/lcd bindings. (**Breaking Change**) (#81)
+- Re-exported printing macros from `pros::io`. (#82)
+- Applied several lints to improve code quality. (#70)
 - Removed the confusingly named `write`, `ewrite`, `writeln`, and `ewriteln` macros. (**Breaking Change**) (#82)
 
 ## [0.7.0]